requirements.txt
scipy==0.19.0
numpy==1.12.0
matplotlib==1.5.3
requests==2.13.0
beautifulsoup4==4.5.3
wikipedia==1.4.0
flask==0.12
flask_bower==1.2.1
tensorflow-gpu==1.0.1
editdistance==0.3.1
sklearn
